third_party/blink/web_tests/platform/mac-mac10.13/virtual/plz-dedicated-worker/external/wpt/service-workers/service-worker/navigation-timing-extended.https-expected.txt

This is a testharness.js-based test.
FAIL Service worker controlled navigation timing assert_greater_than: workerStart marking should not wait for worker activation to finish expected a number greater than 1634042822263.3 but got 1634042822262.5999
Harness: the test ran to completion.
